Rename from fixnum.rb to integer.rb
