+++ lib/v8.rb
@@ -18,7 +18,7 @@ require 'v8/conversion/proc'
 require 'v8/conversion/method'
 require 'v8/conversion/symbol'
 require 'v8/conversion/string'
-require 'v8/conversion/fixnum'
+require 'v8/conversion/integer'
 require 'v8/conversion'
 require 'v8/access/names'
 require 'v8/access/indices'
